refactor: use size_t for local_A and local_C allocations

+++ HW4/part2/matmul.cc
@@ -72,8 +72,8 @@ void matrix_multiply(
 
     const int local_rows = rows[world_rank];
 
-    int *local_A = new int[local_rows * m];
-    int *local_C = new int[local_rows * l];
+    int *local_A = new int[(size_t)local_rows * (size_t)m];
+    int *local_C = new int[(size_t)local_rows * (size_t)l];
 
     MPI_Scatterv(a_mat, numbers_A, offsets_A, MPI_INT, local_A, numbers_A[world_rank], MPI_INT, 0,
                  MPI_COMM_WORLD);
